+++ dbms/src/Functions/arrayIndex.h
@@ -1,27 +1,7 @@
-#pragma once
-
-#include <Common/FieldVisitors.h>
-
-#include <DataTypes/DataTypeArray.h>
-#include <DataTypes/DataTypesNumber.h>
-#include <DataTypes/DataTypeDate.h>
-#include <DataTypes/DataTypeDateTime.h>
-#include <DataTypes/DataTypeString.h>
-#include <DataTypes/DataTypeNullable.h>
-
-#include <Columns/ColumnArray.h>
-#include <Columns/ColumnString.h>
-#include <Columns/ColumnConst.h>
-#include <Columns/ColumnNullable.h>
-
 #include <Functions/IFunction.h>
-#include <Functions/FunctionHelpers.h>
-#include <Common/StringUtils/StringUtils.h>
-
-#include <ext/range.h>
-
-#include <unordered_map>
-#include <numeric>
+#include <Functions/FunctionFactory.h>
+#include <DataTypes/DataTypeArray.h>
+#include <Columns/ColumnArray.h>
 
 
 namespace DB
@@ -30,140 +10,9 @@ namespace DB
 namespace ErrorCodes
 {
     extern const int LOGICAL_ERROR;
+    extern const int ILLEGAL_COLUMN;
+    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
 }
-
-
-/** Array functions:
-  *
-  * array(c1, c2, ...) - create an array.
-  * arrayElement(arr, i) - get the array element by index. If index is not constant and out of range - return default value of data type.
-  * The index begins with 1. Also, the index can be negative - then it is counted from the end of the array.
-  * has(arr, x) - whether there is an element x in the array.
-  * indexOf(arr, x) - returns the index of the element x (starting with 1), if it exists in the array, or 0 if it is not.
-  * arrayEnumerate(arr) - Returns the array [1,2,3,..., length(arr)]
-  *
-  * arrayUniq(arr) - counts the number of different elements in the array,
-  * arrayUniq(arr1, arr2, ...) - counts the number of different tuples from the elements in the corresponding positions in several arrays.
-  *
-  * arrayDistinct(arr) - retrun different elements in an array
-  *
-  * arrayEnumerateUniq(arr)
-  *  - outputs an array parallel (having same size) to this, where for each element specified
-  *  how many times this element was encountered before (including this element) among elements with the same value.
-  *  For example: arrayEnumerateUniq([10, 20, 10, 30]) = [1, 1, 2, 1]
-  * arrayEnumerateUniq(arr1, arr2...)
-  *  - for tuples from elements in the corresponding positions in several arrays.
-  *
-  * emptyArrayToSingle(arr) - replace empty arrays with arrays of one element with a default value.
-  *
-  * arrayReduce('agg', arr1, ...) - apply the aggregate function `agg` to arrays `arr1...`
-  *  If multiple arrays passed, then elements on corresponding positions are passed as multiple arguments to the aggregate function.
-  *
-  *  arrayConcat(arr1, ...) - concatenate arrays.
-  *
-  *  arraySlice(arr, offset, length) - make slice of array. Offsets and length may be < 0 or Null
-  *   - if offset < 0, indexation from right element
-  *   - if length < 0, length = len(array) - (positive_index(offset) - 1) + length
-  *   indexation:
-  *     [ 1,  2,  3,  4,  5,  6]
-  *     [-6, -5, -4, -3, -2, -1]
-  *   examples:
-  *     arraySlice([1, 2, 3, 4, 5, 6], -4, 2) -> [3, 4]
-  *     arraySlice([1, 2, 3, 4, 5, 6], 2, -1) -> [2, 3, 4, 5] (6 - (2 - 1) + (-1) = 4)
-  *     arraySlice([1, 2, 3, 4, 5, 6], -5, -1) = arraySlice([1, 2, 3, 4, 5, 6], 2, -1) -> [2, 3, 4, 5]
-  *
-  *  arrayPushBack(arr, x), arrayPushFront(arr, x)
-  *  arrayPopBack(arr), arrayPopFront(arr)
-  */
-
-
-class FunctionArray : public IFunction
-{
-public:
-    static constexpr auto name = "array";
-    static FunctionPtr create(const Context & context);
-
-    FunctionArray(const Context & context);
-
-    bool useDefaultImplementationForNulls() const override { return false; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    String getName() const override;
-
-    bool addField(DataTypePtr type_res, const Field & f, Array & arr) const;
-
-private:
-    const Context & context;
-};
-
-namespace ArrayImpl
-{
-    class NullMapBuilder;
-}
-
-class FunctionArrayElement : public IFunction
-{
-public:
-    static constexpr auto name = "arrayElement";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    size_t getNumberOfArguments() const override { return 2; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    void perform(Block & block, const ColumnNumbers & arguments, size_t result,
-                     ArrayImpl::NullMapBuilder & builder, size_t input_rows_count);
-
-    template <typename DataType>
-    bool executeNumberConst(Block & block, const ColumnNumbers & arguments, size_t result, const Field & index,
-        ArrayImpl::NullMapBuilder & builder);
-
-    template <typename IndexType, typename DataType>
-    bool executeNumber(Block & block, const ColumnNumbers & arguments, size_t result, const PaddedPODArray<IndexType> & indices,
-        ArrayImpl::NullMapBuilder & builder);
-
-    bool executeStringConst(Block & block, const ColumnNumbers & arguments, size_t result, const Field & index,
-        ArrayImpl::NullMapBuilder & builder);
-
-    template <typename IndexType>
-    bool executeString(Block & block, const ColumnNumbers & arguments, size_t result, const PaddedPODArray<IndexType> & indices,
-        ArrayImpl::NullMapBuilder & builder);
-
-    bool executeGenericConst(Block & block, const ColumnNumbers & arguments, size_t result, const Field & index,
-        ArrayImpl::NullMapBuilder & builder);
-
-    template <typename IndexType>
-    bool executeGeneric(Block & block, const ColumnNumbers & arguments, size_t result, const PaddedPODArray<IndexType> & indices,
-        ArrayImpl::NullMapBuilder & builder);
-
-    template <typename IndexType>
-    bool executeConst(Block & block, const ColumnNumbers & arguments, size_t result,
-                          const PaddedPODArray <IndexType> & indices, ArrayImpl::NullMapBuilder & builder,
-                          size_t input_rows_count);
-
-    template <typename IndexType>
-    bool executeArgument(Block & block, const ColumnNumbers & arguments, size_t result,
-                             ArrayImpl::NullMapBuilder & builder, size_t input_rows_count);
-
-    /** For a tuple array, the function is evaluated component-wise for each element of the tuple.
-      */
-    bool executeTuple(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count);
-};
-
 
 /// For has.
 struct IndexToOne
@@ -1158,588 +1007,5 @@ private:
                 + " of first argument of function " + getName(), ErrorCodes::ILLEGAL_COLUMN};
     }
 };
-
-
-class FunctionArrayEnumerate : public IFunction
-{
-public:
-    static constexpr auto name = "arrayEnumerate";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    size_t getNumberOfArguments() const override { return 1; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-};
-
-
-/// Counts the number of different elements in the array, or the number of different tuples from the elements at the corresponding positions in several arrays.
-/// NOTE The implementation partially matches arrayEnumerateUniq.
-class FunctionArrayUniq : public IFunction
-{
-public:
-    static constexpr auto name = "arrayUniq";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    /// Initially allocate a piece of memory for 512 elements. NOTE: This is just a guess.
-    static constexpr size_t INITIAL_SIZE_DEGREE = 9;
-
-    template <typename T>
-    bool executeNumber(const ColumnArray * array,  const IColumn * null_map, ColumnUInt32::Container & res_values);
-
-    bool executeString(const ColumnArray * array,  const IColumn * null_map, ColumnUInt32::Container & res_values);
-
-    bool execute128bit(
-        const ColumnArray::Offsets & offsets,
-        const ColumnRawPtrs & columns,
-        const ColumnRawPtrs & null_maps,
-        ColumnUInt32::Container & res_values,
-        bool has_nullable_columns);
-
-    void executeHashed(
-        const ColumnArray::Offsets & offsets,
-        const ColumnRawPtrs & columns,
-        ColumnUInt32::Container & res_values);
-};
-
-
-/// Find different elements in an array.
-class FunctionArrayDistinct : public IFunction
-{
-public:
-    static constexpr auto name = "arrayDistinct";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool isVariadic() const override { return false; }
-
-    size_t getNumberOfArguments() const override { return 1; }
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    /// Initially allocate a piece of memory for 512 elements. NOTE: This is just a guess.
-    static constexpr size_t INITIAL_SIZE_DEGREE = 9;
-
-    template <typename T>
-    bool executeNumber(
-        const IColumn & src_data,
-        const ColumnArray::Offsets & src_offsets,
-        IColumn & res_data_col,
-        ColumnArray::Offsets & res_offsets,
-        const ColumnNullable * nullable_col);
-
-    bool executeString(
-        const IColumn & src_data,
-        const ColumnArray::Offsets & src_offsets,
-        IColumn & res_data_col,
-        ColumnArray::Offsets & res_offsets,
-        const ColumnNullable * nullable_col);
-
-    void executeHashed(
-        const IColumn & src_data,
-        const ColumnArray::Offsets & src_offsets,
-        IColumn & res_data_col,
-        ColumnArray::Offsets & res_offsets,
-        const ColumnNullable * nullable_col);
-};
-
-
-template <typename Derived>
-class FunctionArrayEnumerateExtended : public IFunction
-{
-public:
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    /// Initially allocate a piece of memory for 512 elements. NOTE: This is just a guess.
-    static constexpr size_t INITIAL_SIZE_DEGREE = 9;
-
-    template <typename T>
-    bool executeNumber(const ColumnArray * array, const IColumn * null_map, ColumnUInt32::Container & res_values);
-
-    bool executeString(const ColumnArray * array, const IColumn * null_map, ColumnUInt32::Container & res_values);
-
-    bool execute128bit(
-        const ColumnArray::Offsets & offsets,
-        const ColumnRawPtrs & columns,
-        const ColumnRawPtrs & null_maps,
-        ColumnUInt32::Container & res_values,
-        bool has_nullable_columns);
-
-    void executeHashed(
-        const ColumnArray::Offsets & offsets,
-        const ColumnRawPtrs & columns,
-        ColumnUInt32::Container & res_values);
-};
-
-class FunctionArrayEnumerateUniq : public FunctionArrayEnumerateExtended<FunctionArrayEnumerateUniq>
-{
-    using Base = FunctionArrayEnumerateExtended<FunctionArrayEnumerateUniq>;
-public:
-    static constexpr auto name = "arrayEnumerateUniq";
-    using Base::create;
-};
-
-class FunctionArrayEnumerateDense : public FunctionArrayEnumerateExtended<FunctionArrayEnumerateDense>
-{
-    using Base = FunctionArrayEnumerateExtended<FunctionArrayEnumerateDense>;
-public:
-    static constexpr auto name = "arrayEnumerateDense";
-    using Base::create;
-};
-
-template <typename Type> struct TypeToColumnType { using ColumnType = ColumnVector<Type>; };
-template <> struct TypeToColumnType<String> { using ColumnType = ColumnString; };
-
-template <typename DataType> struct DataTypeToName : TypeName<typename DataType::FieldType> { };
-template <> struct DataTypeToName<DataTypeDate> { static std::string get() { return "Date"; } };
-template <> struct DataTypeToName<DataTypeDateTime> { static std::string get() { return "DateTime"; } };
-
-template <typename DataType>
-struct FunctionEmptyArray : public IFunction
-{
-    static constexpr auto base_name = "emptyArray";
-    static const String name;
-    static FunctionPtr create(const Context &) { return std::make_shared<FunctionEmptyArray>(); }
-
-private:
-    String getName() const override
-    {
-        return name;
-    }
-
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & /*arguments*/) const override
-    {
-        return std::make_shared<DataTypeArray>(std::make_shared<DataType>());
-    }
-
-    void executeImpl(Block & block, const ColumnNumbers &, size_t result, size_t input_rows_count) override
-    {
-        using UnderlyingColumnType = typename TypeToColumnType<typename DataType::FieldType>::ColumnType;
-
-        block.getByPosition(result).column = ColumnArray::create(
-            UnderlyingColumnType::create(),
-            ColumnArray::ColumnOffsets::create(input_rows_count, 0));
-    }
-};
-
-template <typename DataType>
-const String FunctionEmptyArray<DataType>::name = FunctionEmptyArray::base_name + String(DataTypeToName<DataType>::get());
-
-class FunctionRange : public IFunction
-{
-public:
-    static constexpr auto name = "range";
-    static FunctionPtr create(const Context &) { return std::make_shared<FunctionRange>(); }
-
-private:
-    String getName() const override;
-
-    size_t getNumberOfArguments() const override { return 1; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    template <typename T>
-    bool executeInternal(Block & block, const IColumn * arg, const size_t result);
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-};
-
-
-class FunctionEmptyArrayToSingle : public IFunction
-{
-public:
-    static constexpr auto name = "emptyArrayToSingle";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    size_t getNumberOfArguments() const override { return 1; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-};
-
-
-class FunctionArrayReverse : public IFunction
-{
-public:
-    static constexpr auto name = "arrayReverse";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    size_t getNumberOfArguments() const override { return 1; }
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-private:
-    bool executeConst(Block & block, const ColumnNumbers & arguments, size_t result,
-                          size_t input_rows_count);
-
-    template <typename T>
-    bool executeNumber(
-        const IColumn & src_data, const ColumnArray::Offsets & src_offsets,
-        IColumn & res_data_col,
-        const ColumnNullable * nullable_col,
-        ColumnNullable * nullable_res_col);
-
-    bool executeFixedString(
-        const IColumn & src_data, const ColumnArray::Offsets & src_offsets,
-        IColumn & res_data_col,
-        const ColumnNullable * nullable_col,
-        ColumnNullable * nullable_res_col);
-
-    bool executeString(
-        const IColumn & src_data, const ColumnArray::Offsets & src_array_offsets,
-        IColumn & res_data_col,
-        const ColumnNullable * nullable_col,
-        ColumnNullable * nullable_res_col);
-};
-
-
-class IAggregateFunction;
-using AggregateFunctionPtr = std::shared_ptr<IAggregateFunction>;
-
-/** Applies an aggregate function to array and returns its result.
-  * If aggregate function has multiple arguments, then this function can be applied to multiple arrays of the same size.
-  */
-class FunctionArrayReduce : public IFunction
-{
-public:
-    static constexpr auto name = "arrayReduce";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    ColumnNumbers getArgumentsThatAreAlwaysConstant() const override { return {0}; }
-
-    DataTypePtr getReturnTypeImpl(const ColumnsWithTypeAndName & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-private:
-    /// lazy initialization in getReturnTypeImpl
-    /// TODO: init in FunctionBuilder
-    mutable AggregateFunctionPtr aggregate_function;
-};
-
-
-class FunctionArrayConcat : public IFunction
-{
-public:
-    static constexpr auto name = "arrayConcat";
-    static FunctionPtr create(const Context & context);
-    FunctionArrayConcat(const Context & context) : context(context) {}
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-private:
-    const Context & context;
-};
-
-
-class FunctionArraySlice : public IFunction
-{
-public:
-    static constexpr auto name = "arraySlice";
-    static FunctionPtr create(const Context & context);
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    bool useDefaultImplementationForNulls() const override { return false; }
-};
-
-
-class FunctionArrayPush : public IFunction
-{
-public:
-    FunctionArrayPush(const Context & context, bool push_front, const char * name)
-        : context(context), push_front(push_front), name(name) {}
-
-    String getName() const override { return name; }
-
-    bool isVariadic() const override { return false; }
-    size_t getNumberOfArguments() const override { return 2; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    bool useDefaultImplementationForNulls() const override { return false; }
-
-private:
-    const Context & context;
-    bool push_front;
-    const char * name;
-};
-
-class FunctionArrayPushFront : public FunctionArrayPush
-{
-public:
-    static constexpr auto name = "arrayPushFront";
-
-    static FunctionPtr create(const Context & context);
-    FunctionArrayPushFront(const Context & context) : FunctionArrayPush(context, true, name) {}
-};
-
-class FunctionArrayPushBack : public FunctionArrayPush
-{
-public:
-    static constexpr auto name = "arrayPushBack";
-
-    static FunctionPtr create(const Context & context);
-    FunctionArrayPushBack(const Context & context) : FunctionArrayPush(context, false, name) {}
-};
-
-class FunctionArrayPop : public IFunction
-{
-public:
-    FunctionArrayPop(bool pop_front, const char * name) : pop_front(pop_front), name(name) {}
-
-    String getName() const override { return name; }
-
-    bool isVariadic() const override { return false; }
-    size_t getNumberOfArguments() const override { return 1; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    bool useDefaultImplementationForNulls() const override { return false; }
-
-private:
-    bool pop_front;
-    const char * name;
-};
-
-class FunctionArrayPopFront : public FunctionArrayPop
-{
-public:
-    static constexpr auto name = "arrayPopFront";
-
-    static FunctionPtr create(const Context & context);
-
-    FunctionArrayPopFront() : FunctionArrayPop(true, name) {}
-};
-
-class FunctionArrayPopBack : public FunctionArrayPop
-{
-public:
-    static constexpr auto name = "arrayPopBack";
-
-    static FunctionPtr create(const Context & context);
-
-    FunctionArrayPopBack() : FunctionArrayPop(false, name) {}
-};
-
-class FunctionArrayIntersect : public IFunction
-{
-public:
-    static constexpr auto name = "arrayIntersect";
-    static FunctionPtr create(const Context & context);
-    FunctionArrayIntersect(const Context & context) : context(context) {}
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-private:
-    const Context & context;
-
-    /// Initially allocate a piece of memory for 512 elements. NOTE: This is just a guess.
-    static constexpr size_t INITIAL_SIZE_DEGREE = 9;
-
-    struct UnpackedArrays
-    {
-        std::vector<char> is_const;
-        std::vector<const NullMap *> null_maps;
-        std::vector<const ColumnArray::ColumnOffsets::Container *> offsets;
-        ColumnRawPtrs nested_columns;
-
-        UnpackedArrays() = default;
-    };
-
-    /// Cast column to data_type removing nullable if data_type hasn't.
-    /// It's expected that column can represent data_type after removing some NullMap's.
-    ColumnPtr castRemoveNullable(const ColumnPtr & column, const DataTypePtr & data_type) const;
-    Columns castColumns(Block & block, const ColumnNumbers & arguments,
-                        const DataTypePtr & return_type, const DataTypePtr & return_type_with_nulls) const;
-    UnpackedArrays prepareArrays(const Columns & columns) const;
-
-    template <typename Map, typename ColumnType, bool is_numeric_column>
-    static ColumnPtr execute(const UnpackedArrays & arrays, MutableColumnPtr result_data);
-
-    struct NumberExecutor
-    {
-        const UnpackedArrays & arrays;
-        const DataTypePtr & data_type;
-        ColumnPtr & result;
-
-        NumberExecutor(const UnpackedArrays & arrays, const DataTypePtr & data_type, ColumnPtr & result)
-            : arrays(arrays), data_type(data_type), result(result) {}
-
-        template <typename T, size_t>
-        void operator()();
-    };
-};
-
-class FunctionArrayHasAllAny : public IFunction
-{
-public:
-    FunctionArrayHasAllAny(const Context & context, bool all, const char * name)
-        : context(context), all(all), name(name) {}
-
-    String getName() const override { return name; }
-
-    bool isVariadic() const override { return false; }
-    size_t getNumberOfArguments() const override { return 2; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-private:
-    const Context & context;
-    bool all;
-    const char * name;
-};
-
-class FunctionArrayHasAll : public FunctionArrayHasAllAny
-{
-public:
-    static constexpr auto name = "hasAll";
-
-    static FunctionPtr create(const Context & context);
-
-    FunctionArrayHasAll(const Context & context) : FunctionArrayHasAllAny(context, true, name) {}
-};
-
-class FunctionArrayHasAny : public FunctionArrayHasAllAny
-{
-public:
-    static constexpr auto name = "hasAny";
-
-    static FunctionPtr create(const Context & context);
-
-    FunctionArrayHasAny(const Context & context) : FunctionArrayHasAllAny(context, false, name) {}
-};
-
-
-class FunctionArrayResize : public IFunction
-{
-public:
-    static constexpr auto name = "arrayResize";
-    static FunctionPtr create(const Context & context);
-    FunctionArrayResize(const Context & context) : context(context) {}
-
-    String getName() const override;
-
-    bool isVariadic() const override { return true; }
-    size_t getNumberOfArguments() const override { return 0; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override;
-
-    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override;
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-    bool useDefaultImplementationForNulls() const override { return false; }
-
-private:
-    const Context & context;
-};
-
-struct NameHas { static constexpr auto name = "has"; };
-struct NameIndexOf { static constexpr auto name = "indexOf"; };
-struct NameCountEqual { static constexpr auto name = "countEqual"; };
-
-using FunctionHas = FunctionArrayIndex<IndexToOne, NameHas>;
-using FunctionIndexOf = FunctionArrayIndex<IndexIdentity, NameIndexOf>;
-using FunctionCountEqual = FunctionArrayIndex<IndexCount, NameCountEqual>;
-
-using FunctionEmptyArrayUInt8 = FunctionEmptyArray<DataTypeUInt8>;
-using FunctionEmptyArrayUInt16 = FunctionEmptyArray<DataTypeUInt16>;
-using FunctionEmptyArrayUInt32 = FunctionEmptyArray<DataTypeUInt32>;
-using FunctionEmptyArrayUInt64 = FunctionEmptyArray<DataTypeUInt64>;
-using FunctionEmptyArrayInt8 = FunctionEmptyArray<DataTypeInt8>;
-using FunctionEmptyArrayInt16 = FunctionEmptyArray<DataTypeInt16>;
-using FunctionEmptyArrayInt32 = FunctionEmptyArray<DataTypeInt32>;
-using FunctionEmptyArrayInt64 = FunctionEmptyArray<DataTypeInt64>;
-using FunctionEmptyArrayFloat32 = FunctionEmptyArray<DataTypeFloat32>;
-using FunctionEmptyArrayFloat64 = FunctionEmptyArray<DataTypeFloat64>;
-using FunctionEmptyArrayDate = FunctionEmptyArray<DataTypeDate>;
-using FunctionEmptyArrayDateTime = FunctionEmptyArray<DataTypeDateTime>;
-using FunctionEmptyArrayString = FunctionEmptyArray<DataTypeString>;
-
 
 }
