+++ dbms/src/Functions/hasAll.cpp
@@ -1,0 +1,21 @@
+#include <Functions/hasAllAny.h>
+#include <Functions/FunctionFactory.h>
+
+
+namespace DB
+{
+
+class FunctionArrayHasAll : public FunctionArrayHasAllAny
+{
+public:
+    static constexpr auto name = "hasAll";
+    static FunctionPtr create(const Context & context) { return std::make_shared<FunctionArrayHasAll>(context); }
+    FunctionArrayHasAll(const Context & context) : FunctionArrayHasAllAny(context, true, name) {}
+};
+
+void registerFunctionArrayHasAll(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionArrayHasAll>();
+}
+
+}
